General fixes PR comments
